refactor: Replace bare `src` package with top-level actor_beautifulsoup_scraper
Give the Actor a descriptively named top-level package instead of one
literally called `src`, run as `python -m actor_beautifulsoup_scraper`.
Repoints the absolute import and updates the Dockerfile and ruff/ty paths.
